requirements.txt
numpy`np
pillow
matplotlib
scipy
